refactor: group render infrastructure
