+++ src/webview/render/registry.ts
