+++ src/webview/sidebar/struct/index.ts
@@ -5,7 +5,7 @@ Pure codec logic lives in struct-codec.ts. */
 
 import { S }        from '../../state';
 import { esc, actionBtnsHtml, wireActionBtns, formatDecimal, formatHex, formatHexHtml, getBigUint64, getBigInt64, asUint64, positionContextMenu, wireHoverSubmenus } from '../../utils';
-import { rerender } from '../../renderRegistry';
+import { rerender } from '../../render/registry';
 import { getByte }  from '../../memory/memoryData';
 import { persistStructPins, persistStructs, persistStructState } from './structPersistence';
 import {
